server/requirements.txt
aerich==0.6.3
aipo[redis,uvloop]==0.1.2
anyio==3.6.1
asyncpg==0.26.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
dictdiffer==0.9.0
eventlet==0.33.2
fastapi==0.79.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
jinja2==3.1.2
pycparser==2.21
pydantic==1.9.1
pypika-tortoise==0.1.6
pytz==2022.1
sniffio==1.2.0
starlette==0.19.1
tomlkit==0.11.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.2
